Removing tests from base group broker mapper test classes
